requirements.txt
streamlit==1.40.0
openai==1.55.3
google-generativeai==0.8.3
PyPDF2==3.0.1
httpx==0.27.2
